ZNTA-1705 - move class so that it's not under webtrans.server package
